Make error messages clearer, disallow empty table in inputs.
This is intended to address the common class of errors I see where
people make a mistake connecting up their modules, but the error message
is either unclear, or doesn't point towards where the mistake actually
is.

The 'what is this in the input' is now explicit about what the problem
is, and if people pass in a nn.Module (meaning they probably forgot a
set of parentheses) instead of a nngraph.Node, we say this explicitly.

The '1 of split(2) outputs unused' (which previously provided no
information about which split was incorrect) now includes file / line
number of both the place where the Node was constructed, and the place
where :split() was called. Hopefully this should reduce debugging time
drastically.

Finally, I have disallow passing an empty table as the input
connections, ie 'nn.Identity()({})' will error. I cannot see a use case
for this (if you have no input connections, just leave the second parens
empty). The risk of this is when people do
'nn.Identity()({variableWithTypo})', thinking they have made a
connection but actually they haven't. This is likely to cause errors
much later on, whereas with this commit it errors straight away.

This *could* break existing code, but theres an easy to apply fix that
needs to be done at each callsite. Koray has approved this restriction
to the API, but I appreciate others may have a view here..