Store a reverse mapping when wiring together graph, detect unused nodes.
The connectivity checking code was previously unable to detect the
following error case:

  local input = nn.Identity()()
  local usedOutput = nn.Linear(20, 10)(input)
  local unusedOutput = nn.Linear(20, 10)(input)
  local gmod = nn.gModule({input}, {usedOutput})

With this fix, when gModule is called it will throw an error, because of
unusedOutput.

This is a backwards incompatible change, but I feel that the current
flexibility is error prone, and I can't see any advantage to it. We have
flushed out a couple of bugs in internal code with this change.